rename stamp to achievement in test code
