+++ app-android/src/test/java/io/github/droidkaigi/confsched2023/KaigiAppTest.kt
@@ -82,9 +82,9 @@ class KaigiAppTest {
     }
 
     @Test
-    fun checkNavigateToStampsShot() {
+    fun checkNavigateToAchievementsShot() {
         kaigiAppRobot {
-            goToStamps()
+            goToAchievements()
             capture()
         }
     }
