+++ core/testing/src/main/java/io/github/droidkaigi/confsched2023/testing/robot/KaigiAppRobot.kt
@@ -52,9 +52,9 @@ class KaigiAppRobot @Inject constructor(
         waitUntilIdle()
     }
 
-    fun goToStamps() {
+    fun goToAchievements() {
         composeTestRule
-            .onNode(hasTestTag(MainScreenTab.Badges.testTag))
+            .onNode(hasTestTag(MainScreenTab.Achievements.testTag))
             .performClick()
         waitUntilIdle()
     }
