+++ feature/achievements/src/test/java/io/github/droidkaigi/confsched2023/achievements/AchievementsScreenTest.kt
@@ -1,4 +1,4 @@
-package io.github.droidkaigi.confsched2023.stamps
+package io.github.droidkaigi.confsched2023.achievements
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.github.takahirom.roborazzi.RobolectricDeviceQualifiers
@@ -7,7 +7,7 @@ import dagger.hilt.android.testing.HiltAndroidTest
 import io.github.droidkaigi.confsched2023.testing.HiltTestActivity
 import io.github.droidkaigi.confsched2023.testing.RobotTestRule
 import io.github.droidkaigi.confsched2023.testing.category.ScreenshotTests
-import io.github.droidkaigi.confsched2023.testing.robot.StampsScreenRobot
+import io.github.droidkaigi.confsched2023.testing.robot.AchievementsScreenRobot
 import org.junit.Rule
 import org.junit.Test
 import org.junit.experimental.categories.Category
@@ -22,18 +22,18 @@ import javax.inject.Inject
 @Config(
     qualifiers = RobolectricDeviceQualifiers.NexusOne,
 )
-class StampsScreenTest {
+class AchievementsScreenTest {
 
     @get:Rule
     @BindValue val robotTestRule: RobotTestRule = RobotTestRule<HiltTestActivity>(this)
 
     @Inject
-    lateinit var stampsScreenRobot: StampsScreenRobot
+    lateinit var achievementsScreenRobot: AchievementsScreenRobot
 
     @Test
     @Category(ScreenshotTests::class)
     fun checkLaunchShot() {
-        stampsScreenRobot {
+        achievementsScreenRobot {
             setupScreenContent()
             checkScreenCapture()
             scroll()
